requirements.txt
boto3
requests
requests-aws4auth
opensearch-py


azure-ai-documentintelligence
azure-core
pandas
matplotlib
seaborn